refactor: fixes after code-review
